refactoring: renaming PieceAbstraction to PieceModel
